Move EulaButtonElement to correct package
