+++ skin-manager-shared/src/main/java/de/hdskins/labymod/shared/settings/element/DefaultElementFactory.java
@@ -21,7 +21,7 @@ import de.hdskins.labymod.shared.settings.element.elements.ButtonElement;
 import de.hdskins.labymod.shared.settings.element.elements.ChangeableBooleanElement;
 import de.hdskins.labymod.shared.settings.element.elements.CustomDropDownElement;
 import de.hdskins.labymod.shared.settings.element.elements.PlayerSkinRenderElement;
-import de.hdskins.labymod.shared.settings.eula.EulaButtonElement;
+import de.hdskins.labymod.shared.settings.element.elements.EulaButtonElement;
 import net.labymod.settings.elements.ControlElement;
 import net.labymod.settings.elements.DropDownElement;
 import net.labymod.utils.Consumer;
